Add extrinsic imu calibration
