+++ bitbots_misc/bitbots_extrinsic_calibration/config/default.yaml
@@ -1,4 +1,10 @@
-/bitbots_extrinsic_calibration:
+/bitbots_extrinsic_camera_calibration:
+  ros__parameters:
+    offset_x: 0.0
+    offset_y: 0.0
+    offset_z: 0.0
+
+/bitbots_extrinsic_imu_calibration:
   ros__parameters:
     offset_x: 0.0
     offset_y: 0.0
